Refactored main frame
