+++ isolines/src/main/java/ru/nsu/fit/semenov/isolines/Main.java
@@ -1,5 +1,7 @@
 package ru.nsu.fit.semenov.isolines;
 
+import ru.nsu.fit.semenov.isolines.my_frames.MainFrame;
+
 public class Main {
 
     public static void main(String[] args) {
