+++ isolines/src/main/java/ru/nsu/fit/semenov/isolines/model/BoundedFunction.java
@@ -1,4 +1,4 @@
-package ru.nsu.fit.semenov.isolines;
+package ru.nsu.fit.semenov.isolines.model;
 
 public interface BoundedFunction {
