+++ isolines/src/main/java/ru/nsu/fit/semenov/isolines/model/BoundedFunctionImpl.java
@@ -1,4 +1,4 @@
-package ru.nsu.fit.semenov.isolines;
+package ru.nsu.fit.semenov.isolines.model;
 
 public class BoundedFunctionImpl implements BoundedFunction {
