+++ isolines/src/main/java/ru/nsu/fit/semenov/isolines/my_frames/ImageLayerDataIndex.java
@@ -1,0 +1,11 @@
+package ru.nsu.fit.semenov.isolines.my_frames;
+
+public enum ImageLayerDataIndex {
+
+    MAP,
+    INTERPOLATED_MAP,
+    ISOLINES,
+    GRID,
+    CONTROL_POINTS
+
+}
